+++ db/migrate/20220420020457_add_open_in_new_tab_to_admin_resources.rb
@@ -1,0 +1,5 @@
+class AddOpenInNewTabToAdminResources < ActiveRecord::Migration[6.1]
+  def change
+    add_column :admin_resources, :open_in_new_tab, :boolean, null: false, default: false
+  end
+end